requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Authlib==1.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.1.1
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
drf-jwt==1.13.4
drf-yasg==1.21.3
flower==1.2.0
humanize==4.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
